+++ src/test/kotlin/proguard/MultiTypeTest.kt
@@ -23,6 +23,7 @@ import proguard.evaluation.BasicInvocationUnit
 import proguard.evaluation.PartialEvaluator
 import proguard.evaluation.value.MultiTypedReferenceValue
 import proguard.evaluation.value.MultiTypedReferenceValueFactory
+import proguard.evaluation.value.ParticularIntegerValue
 import proguard.evaluation.value.TypedReferenceValue
 import proguard.evaluation.value.UnknownReferenceValue
 import proguard.evaluation.value.Value
@@ -90,12 +91,17 @@ class MultiTypeTest : FreeSpec({
                     Super s = new Super();
                     // s is always Super
                 }
+                
+                public void array() {
+                    A[] array = new A[10];
+                    A a = array[3];
+                }
             }
         """
     val classPool = ClassPoolBuilder.fromStrings(listOf("-g"), codeSuper, codeA, codeB, codeTarget)
 
     "Code examples" - {
-        "Exact type" - {
+        "Exact type" {
             val (instructions, variableTable) = evaluate(
                 "Target",
                 "exact",
@@ -110,7 +116,7 @@ class MultiTypeTest : FreeSpec({
             s.generalizedType.type shouldBe "Super"
             s.potentialTypes.map { it.type } shouldBe listOf("Super")
         }
-        "Ternary operator" - {
+        "Ternary operator" {
             val (instructions, variableTable) = evaluate(
                 "Target",
                 "ternary",
@@ -126,7 +132,7 @@ class MultiTypeTest : FreeSpec({
             s.isNull shouldBe Value.NEVER
             s.potentialTypes.map { it.type }.toSet() shouldBe setOf("A", "B")
         }
-        "If else" - {
+        "If else" {
             val (instructions, variableTable) = evaluate(
                 "Target",
                 "ifElse",
@@ -142,7 +148,7 @@ class MultiTypeTest : FreeSpec({
             s.isNull shouldBe Value.NEVER
             s.potentialTypes.map { it.type }.toSet() shouldBe setOf("A", "B")
         }
-        "Switch" - {
+        "Switch" {
             val (instructions, variableTable) = evaluate(
                 "Target",
                 "switchStmt",
@@ -157,6 +163,28 @@ class MultiTypeTest : FreeSpec({
             s.generalizedType.type shouldBe "Super"
             s.isNull shouldBe Value.MAYBE
             s.potentialTypes.map { it.type }.toSet() shouldBe setOf("A", "B", "Super", null)
+        }
+        "Array handling" {
+            val (instructions, variableTable) = evaluate(
+                "Target",
+                "array",
+                "()V",
+                classPool,
+                partialEvaluator
+            )
+            val (methodEnd, _) = instructions.last()
+
+            val array = partialEvaluator.getVariablesBefore(methodEnd)
+                .getValue(variableTable["array"]!!) as MultiTypedReferenceValue
+            array.generalizedType.type shouldBe "[LA;"
+            array.isNull shouldBe Value.NEVER
+            array.potentialTypes.map { it.type }.toSet() shouldBe setOf("[LA;")
+
+            val a = partialEvaluator.getVariablesBefore(methodEnd)
+                .getValue(variableTable["a"]!!) as MultiTypedReferenceValue
+            a.generalizedType.type shouldBe "A"
+            a.isNull shouldBe Value.MAYBE
+            a.potentialTypes.map { it.type }.toSet() shouldBe setOf("A")
         }
     }
 
@@ -188,14 +216,22 @@ class MultiTypeTest : FreeSpec({
         )
         val multiB = MultiTypedReferenceValue(b, false)
 
+        val arrayA = TypedReferenceValue(
+            "[LA;",
+            classPool.getClass("A"),
+            false,
+            false
+        )
+        val multiArrayA = MultiTypedReferenceValue(arrayA, false)
+
         "Generalize" - {
-            "(A, B) -> Super" - {
+            "(A, B) -> Super" {
                 val generalized = multiA.generalize(multiB) as MultiTypedReferenceValue
                 generalized.generalizedType shouldBe superClass
                 generalized.potentialTypes shouldBe setOf(a, b)
                 generalized.mayBeUnknown shouldBe false
             }
-            "(X, Super) -> Super" - {
+            "(X, Super) -> Super" {
                 var generalized = multiA.generalize(multiSuper) as MultiTypedReferenceValue
                 generalized.generalizedType shouldBe superClass
                 generalized.potentialTypes shouldBe setOf(a, superClass)
@@ -206,7 +242,7 @@ class MultiTypeTest : FreeSpec({
                 generalized.potentialTypes shouldBe setOf(b, superClass)
                 generalized.mayBeUnknown shouldBe false
             }
-            "(X, null) -> X" - {
+            "(X, null) -> X" {
                 var generalized = multiA.generalize(multiNull) as MultiTypedReferenceValue
                 generalized.generalizedType shouldBe a
                 generalized.potentialTypes shouldBe setOf(a, nul)
@@ -222,7 +258,7 @@ class MultiTypeTest : FreeSpec({
                 generalized.potentialTypes shouldBe setOf(superClass, nul)
                 generalized.mayBeUnknown shouldBe false
             }
-            "Identity" - {
+            "Identity" {
                 var generalized = multiA.generalize(multiA) as MultiTypedReferenceValue
                 generalized.generalizedType shouldBe a
                 generalized.potentialTypes shouldBe setOf(a)
@@ -242,8 +278,13 @@ class MultiTypeTest : FreeSpec({
                 generalized.generalizedType shouldBe nul
                 generalized.potentialTypes shouldBe setOf(nul)
                 generalized.mayBeUnknown shouldBe false
+
+                generalized = multiArrayA.generalize(multiArrayA) as MultiTypedReferenceValue
+                generalized.generalizedType shouldBe arrayA
+                generalized.potentialTypes shouldBe setOf(arrayA)
+                generalized.mayBeUnknown shouldBe false
             }
-            "Handle TypedReferenceValues transparently" - {
+            "Handle TypedReferenceValues transparently" {
                 var generalized = multiA.generalize(a) as MultiTypedReferenceValue
                 generalized.generalizedType shouldBe a
                 generalized.potentialTypes shouldBe setOf(a)
@@ -263,8 +304,13 @@ class MultiTypeTest : FreeSpec({
                 generalized.generalizedType shouldBe nul
                 generalized.potentialTypes shouldBe setOf(nul)
                 generalized.mayBeUnknown shouldBe false
+
+                generalized = multiArrayA.generalize(arrayA) as MultiTypedReferenceValue
+                generalized.generalizedType shouldBe arrayA
+                generalized.potentialTypes shouldBe setOf(arrayA)
+                generalized.mayBeUnknown shouldBe false
             }
-            "Handle UnknownReferenceValue" - {
+            "Handle UnknownReferenceValue" {
                 var generalized =
                     multiA.generalize(UnknownReferenceValue()) as MultiTypedReferenceValue
                 generalized.generalizedType shouldBe a
@@ -282,8 +328,26 @@ class MultiTypeTest : FreeSpec({
                 generalized.potentialTypes shouldBe setOf(superClass)
                 generalized.mayBeUnknown shouldBe true
             }
+            "Handle array dereference" {
+                val dereferenced = multiArrayA.referenceArrayLoad(
+                    ParticularIntegerValue(42),
+                    valueFactory
+                ) as MultiTypedReferenceValue
+
+                dereferenced.type shouldBe "A"
+                dereferenced.referencedClass shouldBe classPool.getClass("A")
+
+                val referenced = valueFactory.createArrayReferenceValue(
+                    "LA;",
+                    classPool.getClass("A"),
+                    ParticularIntegerValue(42)
+                ) as MultiTypedReferenceValue
+
+                referenced.type shouldBe "[LA;"
+                referenced.referencedClass shouldBe classPool.getClass("A")
+            }
         }
-        "Null check" - {
+        "Null check" {
             multiNull.isNull shouldBe Value.ALWAYS
 
             multiA.isNull shouldBe Value.NEVER
@@ -294,7 +358,7 @@ class MultiTypeTest : FreeSpec({
             multiA.generalize(multiNull).isNull shouldBe Value.MAYBE
         }
         "Casts" - {
-            "Upcasting preserves type" - {
+            "Upcasting preserves type" {
                 multiA.cast(
                     superClass.type,
                     superClass.referencedClass,
@@ -308,7 +372,7 @@ class MultiTypeTest : FreeSpec({
                     true
                 ) shouldBe multiB
             }
-            "Downcasting overrides type" - {
+            "Downcasting overrides type" {
                 multiSuper.cast(
                     a.type,
                     a.referencedClass,